+++ ghostwriter/users/admin.py
@@ -42,6 +42,7 @@ class UserAdmin(auth_admin.UserAdmin):
         "is_active",
     )
     list_filter = ("role",)
+    search_fields = ("username", "name", "email")
 
     fieldsets = (
         (_("User Information"), {"fields": ("username", "password", "require_mfa")}),